cvora: add a soft rearm command to libcvora
As reported by Piotr Nikiel <piotr.pawel.nikiel@cern.ch>,
the memory pointer register that reports the size of an acquisiion
reports bogus data with a definite statiscal pattern. This happens
in every CVORA mode.

A software fix for this is to perform a soft rearm command. This
patch adds this function to the library.